Dratt ut sendVedlegg for dok.etterspurt i egen komponent
